EV-AS21: Refactoring of existing methods
