+++ src/rpc_endpoints/signup.rs
@@ -1,6 +1,5 @@
 use super::super::storage_ops::storage_manager::StorageManager;
-use super::account_service::{SignUpRequest, SignUpResponse};
-
+use super::account_service_endpoints::account_service::{SignUpRequest, SignUpResponse};
 use tonic::{Request, Response, Status};
 
 pub mod account_service {
@@ -20,9 +19,6 @@ pub async fn sign_up(
     request: Request<SignUpRequest>,
     storage_manager: &StorageManager,
 ) -> Result<Response<SignUpResponse>, Status> {
-    println!("Username = {}", request.get_ref().username);
-    println!("Password = {}", request.get_ref().password);
-
     storage_manager.get_all_accounts().await;
     let _ = storage_manager.get_account_by_username("vladonis").await;
